DdmDatasetController Changes and Module html changes
Added a Database Handler Class for the DdmDatasetController class to use.
This class is used for all interaction with the database that stores the
information that is related to only this module. The way it is currently
implimented is maybe not the cleanest way to do it but for now it is working.
It currently also extends the Gridsubprocess class and has direct access to
the grid. There might be a better way to impliment this but I was unable to
find one.
Fixed a bug in the html page for the module. Before it was displaying 10 rows
of all the same dataset information. Now it actually displays the 10
different datasets that are retrieved from the controller.