Last Run Time Checking Added
DdmDatasetsController now checks when the last time the run method was called
by write the run time or the run method out to a log file. It then reads from
the log file the last time the run method was called the next time the run
method is executed. If it was executed within a certain amount of time ago,
the run method does nothing. If it was longer than that, then the run method
writes out the current run time and then continues with normal functionality.
One bug currently exists. Last time I ran it there was a permisions error
when trying to write to the log file. Work on this will continue.