Minor Changes
Some minor changes were made in the DdmController and the DdmViewer. In the
controller, some variables were abstracted to be easily change able at the
very top of the module. In the viewier the name of the pie chart was changed
from the default to an applicable title. Also the location of the log file
for the controller has been moved to /var/lib/HappyFace3/log/DdmLastRun.log