Added Config File Support for DDM Module
The DdmDatasetsModule now reads parameters from the config file for the
module and passes those into the constructor of the controller. This allows
for values to be set in the configuration file that affect how the controller
acts. A major one is whether the module should be run in production mode or
in test mode. It is now very easy to switch from a development environment
to a production environment just by chaning this value.